+++ lib/src/chart/bar_chart/bar_chart_data.dart
@@ -48,6 +48,7 @@ class BarChartData extends AxisChartData with EquatableMixin {
     double? groupsSpace,
     BarChartAlignment? alignment,
     FlTitlesData? titlesData,
+    ExtraLinesData? extraLinesData,
     BarTouchData? barTouchData,
     double? maxY,
     double? minY,
@@ -71,6 +72,7 @@ class BarChartData extends AxisChartData with EquatableMixin {
               ),
           gridData: gridData ?? FlGridData(),
           borderData: borderData,
+          extraLinesData: extraLinesData,
           rangeAnnotations: rangeAnnotations ?? RangeAnnotations(),
           backgroundColor: backgroundColor,
           touchData: barTouchData ?? BarTouchData(),
@@ -83,6 +85,30 @@ class BarChartData extends AxisChartData with EquatableMixin {
           baselineY: baselineY,
         );
 
+  /// Lerps a [BaseChartData] based on [t] value, check [Tween.lerp].
+  @override
+  BarChartData lerp(BaseChartData a, BaseChartData b, double t) {
+    if (a is BarChartData && b is BarChartData) {
+      return BarChartData(
+        barGroups: lerpBarChartGroupDataList(a.barGroups, b.barGroups, t),
+        groupsSpace: lerpDouble(a.groupsSpace, b.groupsSpace, t),
+        alignment: b.alignment,
+        titlesData: FlTitlesData.lerp(a.titlesData, b.titlesData, t),
+        extraLinesData: ExtraLinesData.lerp(a.extraLinesData, b.extraLinesData, t),
+        rangeAnnotations: RangeAnnotations.lerp(a.rangeAnnotations, b.rangeAnnotations, t),
+        barTouchData: b.barTouchData,
+        gridData: FlGridData.lerp(a.gridData, b.gridData, t),
+        borderData: FlBorderData.lerp(a.borderData, b.borderData, t),
+        maxY: lerpDouble(a.maxY, b.maxY, t),
+        minY: lerpDouble(a.minY, b.minY, t),
+        baselineY: lerpDouble(a.baselineY, b.baselineY, t),
+        backgroundColor: Color.lerp(a.backgroundColor, b.backgroundColor, t),
+      );
+    } else {
+      throw Exception('Illegal State');
+    }
+  }
+
   /// Copies current [BarChartData] to a new [BarChartData],
   /// and replaces provided values.
   BarChartData copyWith({
@@ -90,6 +116,7 @@ class BarChartData extends AxisChartData with EquatableMixin {
     double? groupsSpace,
     BarChartAlignment? alignment,
     FlTitlesData? titlesData,
+    ExtraLinesData? extraLinesData,
     RangeAnnotations? rangeAnnotations,
     BarTouchData? barTouchData,
     FlGridData? gridData,
@@ -104,6 +131,7 @@ class BarChartData extends AxisChartData with EquatableMixin {
       groupsSpace: groupsSpace ?? this.groupsSpace,
       alignment: alignment ?? this.alignment,
       titlesData: titlesData ?? this.titlesData,
+      extraLinesData: extraLinesData ?? this.extraLinesData,
       rangeAnnotations: rangeAnnotations ?? this.rangeAnnotations,
       barTouchData: barTouchData ?? this.barTouchData,
       gridData: gridData ?? this.gridData,
@@ -115,30 +143,6 @@ class BarChartData extends AxisChartData with EquatableMixin {
     );
   }
 
-  /// Lerps a [BaseChartData] based on [t] value, check [Tween.lerp].
-  @override
-  BarChartData lerp(BaseChartData a, BaseChartData b, double t) {
-    if (a is BarChartData && b is BarChartData) {
-      return BarChartData(
-        barGroups: lerpBarChartGroupDataList(a.barGroups, b.barGroups, t),
-        groupsSpace: lerpDouble(a.groupsSpace, b.groupsSpace, t),
-        alignment: b.alignment,
-        titlesData: FlTitlesData.lerp(a.titlesData, b.titlesData, t),
-        rangeAnnotations:
-            RangeAnnotations.lerp(a.rangeAnnotations, b.rangeAnnotations, t),
-        barTouchData: b.barTouchData,
-        gridData: FlGridData.lerp(a.gridData, b.gridData, t),
-        borderData: FlBorderData.lerp(a.borderData, b.borderData, t),
-        maxY: lerpDouble(a.maxY, b.maxY, t),
-        minY: lerpDouble(a.minY, b.minY, t),
-        baselineY: lerpDouble(a.baselineY, b.baselineY, t),
-        backgroundColor: Color.lerp(a.backgroundColor, b.backgroundColor, t),
-      );
-    } else {
-      throw Exception('Illegal State');
-    }
-  }
-
   /// Used for equality check, see [EquatableMixin].
   @override
   List<Object?> get props => [
@@ -146,6 +150,7 @@ class BarChartData extends AxisChartData with EquatableMixin {
         groupsSpace,
         alignment,
         titlesData,
+        extraLinesData,
         barTouchData,
         maxY,
         minY,
@@ -253,7 +258,7 @@ class BarChartGroupData with EquatableMixin {
   static BarChartGroupData lerp(
       BarChartGroupData a, BarChartGroupData b, double t) {
     return BarChartGroupData(
-      x: (a.x + (b.x - a.x) * t).round(),
+      x: (a.x + (b.x - a.x) * t + 50).round(),
       groupVertically: b.groupVertically,
       barRods: lerpBarChartRodDataList(a.barRods, b.barRods, t),
       barsSpace: lerpDouble(a.barsSpace, b.barsSpace, t),