+++ lib/src/chart/line_chart/line_chart_data.dart
@@ -20,9 +20,6 @@ class LineChartData extends AxisChartData with EquatableMixin {
   /// Fills area between two [LineChartBarData] with a color or gradient.
   final List<BetweenBarsData> betweenBarsData;
 
-  /// [LineChart] draws some horizontal or vertical lines on above or below of everything
-  final ExtraLinesData extraLinesData;
-
   /// Handles touch behaviors and responses.
   final LineTouchData lineTouchData;
 
@@ -78,7 +75,6 @@ class LineChartData extends AxisChartData with EquatableMixin {
     Color? backgroundColor,
   })  : lineBarsData = lineBarsData ?? const [],
         betweenBarsData = betweenBarsData ?? const [],
-        extraLinesData = extraLinesData ?? ExtraLinesData(),
         lineTouchData = lineTouchData ?? LineTouchData(),
         showingTooltipIndicators = showingTooltipIndicators ?? const [],
         super(
@@ -86,22 +82,15 @@ class LineChartData extends AxisChartData with EquatableMixin {
           touchData: lineTouchData ?? LineTouchData(),
           borderData: borderData,
           titlesData: titlesData ?? FlTitlesData(),
+          extraLinesData: extraLinesData,
           rangeAnnotations: rangeAnnotations ?? RangeAnnotations(),
           clipData: clipData ?? FlClipData.none(),
           backgroundColor: backgroundColor,
-          minX: minX ??
-              LineChartHelper.calculateMaxAxisValues(lineBarsData ?? const [])
-                  .minX,
-          maxX: maxX ??
-              LineChartHelper.calculateMaxAxisValues(lineBarsData ?? const [])
-                  .maxX,
+          minX: minX ?? LineChartHelper.calculateMaxAxisValues(lineBarsData ?? const []).minX,
+          maxX: maxX ?? LineChartHelper.calculateMaxAxisValues(lineBarsData ?? const []).maxX,
           baselineX: baselineX,
-          minY: minY ??
-              LineChartHelper.calculateMaxAxisValues(lineBarsData ?? const [])
-                  .minY,
-          maxY: maxY ??
-              LineChartHelper.calculateMaxAxisValues(lineBarsData ?? const [])
-                  .maxY,
+          minY: minY ?? LineChartHelper.calculateMaxAxisValues(lineBarsData ?? const []).minY,
+          maxY: maxY ?? LineChartHelper.calculateMaxAxisValues(lineBarsData ?? const []).maxY,
           baselineY: baselineY,
         );
 
@@ -119,16 +108,12 @@ class LineChartData extends AxisChartData with EquatableMixin {
         backgroundColor: Color.lerp(a.backgroundColor, b.backgroundColor, t),
         borderData: FlBorderData.lerp(a.borderData, b.borderData, t),
         clipData: b.clipData,
-        extraLinesData:
-            ExtraLinesData.lerp(a.extraLinesData, b.extraLinesData, t),
+        extraLinesData: ExtraLinesData.lerp(a.extraLinesData, b.extraLinesData, t),
         gridData: FlGridData.lerp(a.gridData, b.gridData, t),
         titlesData: FlTitlesData.lerp(a.titlesData, b.titlesData, t),
-        rangeAnnotations:
-            RangeAnnotations.lerp(a.rangeAnnotations, b.rangeAnnotations, t),
-        lineBarsData:
-            lerpLineChartBarDataList(a.lineBarsData, b.lineBarsData, t),
-        betweenBarsData:
-            lerpBetweenBarsDataList(a.betweenBarsData, b.betweenBarsData, t),
+        rangeAnnotations: RangeAnnotations.lerp(a.rangeAnnotations, b.rangeAnnotations, t),
+        lineBarsData: lerpLineChartBarDataList(a.lineBarsData, b.lineBarsData, t),
+        betweenBarsData: lerpBetweenBarsDataList(a.betweenBarsData, b.betweenBarsData, t),
         lineTouchData: b.lineTouchData,
         showingTooltipIndicators: b.showingTooltipIndicators,
       );
@@ -165,8 +150,7 @@ class LineChartData extends AxisChartData with EquatableMixin {
       rangeAnnotations: rangeAnnotations ?? this.rangeAnnotations,
       extraLinesData: extraLinesData ?? this.extraLinesData,
       lineTouchData: lineTouchData ?? this.lineTouchData,
-      showingTooltipIndicators:
-          showingTooltipIndicators ?? this.showingTooltipIndicators,
+      showingTooltipIndicators: showingTooltipIndicators ?? this.showingTooltipIndicators,
       gridData: gridData ?? this.gridData,
       borderData: borderData ?? this.borderData,
       minX: minX ?? this.minX,
@@ -346,15 +330,13 @@ class LineChartBarData with EquatableMixin {
     LineChartStepData? lineChartStepData,
   })  : spots = spots ?? const [],
         show = show ?? true,
-        color =
-            color ?? ((color == null && gradient == null) ? Colors.cyan : null),
+        color = color ?? ((color == null && gradient == null) ? Colors.cyan : null),
         gradient = gradient,
         barWidth = barWidth ?? 2.0,
         isCurved = isCurved ?? false,
         curveSmoothness = curveSmoothness ?? 0.35,
         preventCurveOverShooting = preventCurveOverShooting ?? false,
-        preventCurveOvershootingThreshold =
-            preventCurveOvershootingThreshold ?? 10.0,
+        preventCurveOvershootingThreshold = preventCurveOvershootingThreshold ?? 10.0,
         isStrokeCapRound = isStrokeCapRound ?? false,
         isStrokeJoinRound = isStrokeJoinRound ?? false,
         belowBarData = belowBarData ?? BarAreaData(),
@@ -369,8 +351,7 @@ class LineChartBarData with EquatableMixin {
 
     FlSpot? firstValidSpot;
     try {
-      firstValidSpot =
-          this.spots.firstWhere((element) => element != FlSpot.nullSpot);
+      firstValidSpot = this.spots.firstWhere((element) => element != FlSpot.nullSpot);
     } catch (e) {
       // There is no valid spot
     }
@@ -403,8 +384,7 @@ class LineChartBarData with EquatableMixin {
   }
 
   /// Lerps a [LineChartBarData] based on [t] value, check [Tween.lerp].
-  static LineChartBarData lerp(
-      LineChartBarData a, LineChartBarData b, double t) {
+  static LineChartBarData lerp(LineChartBarData a, LineChartBarData b, double t) {
     return LineChartBarData(
       show: b.show,
       barWidth: lerpDouble(a.barWidth, b.barWidth, t),
@@ -415,10 +395,8 @@ class LineChartBarData with EquatableMixin {
       isStrokeCapRound: b.isStrokeCapRound,
       isStrokeJoinRound: b.isStrokeJoinRound,
       preventCurveOverShooting: b.preventCurveOverShooting,
-      preventCurveOvershootingThreshold: lerpDouble(
-          a.preventCurveOvershootingThreshold,
-          b.preventCurveOvershootingThreshold,
-          t),
+      preventCurveOvershootingThreshold:
+          lerpDouble(a.preventCurveOvershootingThreshold, b.preventCurveOvershootingThreshold, t),
       dotData: FlDotData.lerp(a.dotData, b.dotData, t),
       dashArray: lerpIntList(a.dashArray, b.dashArray, t),
       color: Color.lerp(a.color, b.color, t),
@@ -427,8 +405,7 @@ class LineChartBarData with EquatableMixin {
       showingIndicators: b.showingIndicators,
       shadow: Shadow.lerp(a.shadow, b.shadow, t),
       isStepLineChart: b.isStepLineChart,
-      lineChartStepData:
-          LineChartStepData.lerp(a.lineChartStepData, b.lineChartStepData, t),
+      lineChartStepData: LineChartStepData.lerp(a.lineChartStepData, b.lineChartStepData, t),
     );
   }
 
@@ -463,10 +440,8 @@ class LineChartBarData with EquatableMixin {
       barWidth: barWidth ?? this.barWidth,
       isCurved: isCurved ?? this.isCurved,
       curveSmoothness: curveSmoothness ?? this.curveSmoothness,
-      preventCurveOverShooting:
-          preventCurveOverShooting ?? this.preventCurveOverShooting,
-      preventCurveOvershootingThreshold: preventCurveOvershootingThreshold ??
-          this.preventCurveOvershootingThreshold,
+      preventCurveOverShooting: preventCurveOverShooting ?? this.preventCurveOverShooting,
+      preventCurveOvershootingThreshold: preventCurveOvershootingThreshold ?? this.preventCurveOvershootingThreshold,
       isStrokeCapRound: isStrokeCapRound ?? this.isStrokeCapRound,
       isStrokeJoinRound: isStrokeJoinRound ?? this.isStrokeJoinRound,
       belowBarData: belowBarData ?? this.belowBarData,
@@ -523,8 +498,7 @@ class LineChartStepData with EquatableMixin {
   LineChartStepData({this.stepDirection = stepDirectionMiddle});
 
   /// Lerps a [LineChartStepData] based on [t] value, check [Tween.lerp].
-  static LineChartStepData lerp(
-      LineChartStepData a, LineChartStepData b, double t) {
+  static LineChartStepData lerp(LineChartStepData a, LineChartStepData b, double t) {
     return LineChartStepData(
       stepDirection: lerpDouble(a.stepDirection, b.stepDirection, t)!,
     );
@@ -580,10 +554,7 @@ class BarAreaData with EquatableMixin {
     double? cutOffY,
     bool? applyCutOffY,
   })  : show = show ?? false,
-        color = color ??
-            ((color == null && gradient == null)
-                ? Colors.blueGrey.withOpacity(0.5)
-                : null),
+        color = color ?? ((color == null && gradient == null) ? Colors.blueGrey.withOpacity(0.5) : null),
         gradient = gradient,
         spotsLine = spotsLine ?? BarAreaSpotsLine(),
         cutOffY = cutOffY ?? 0,
@@ -640,10 +611,7 @@ class BetweenBarsData with EquatableMixin {
     Gradient? gradient,
   })  : fromIndex = fromIndex,
         toIndex = toIndex,
-        color = color ??
-            ((color == null && gradient == null)
-                ? Colors.blueGrey.withOpacity(0.5)
-                : null),
+        color = color ?? ((color == null && gradient == null) ? Colors.blueGrey.withOpacity(0.5) : null),
         gradient = gradient;
 
   /// Lerps a [BetweenBarsData] based on [t] value, check [Tween.lerp].
@@ -694,8 +662,7 @@ class BarAreaSpotsLine with EquatableMixin {
         applyCutOffY = applyCutOffY ?? true;
 
   /// Lerps a [BarAreaSpotsLine] based on [t] value, check [Tween.lerp].
-  static BarAreaSpotsLine lerp(
-      BarAreaSpotsLine a, BarAreaSpotsLine b, double t) {
+  static BarAreaSpotsLine lerp(BarAreaSpotsLine a, BarAreaSpotsLine b, double t) {
     return BarAreaSpotsLine(
       show: b.show,
       checkToShowSpotLine: b.checkToShowSpotLine,
@@ -748,8 +715,7 @@ Color _defaultGetDotColor(FlSpot _, double xPercentage, LineChartBarData bar) {
 
 /// If there is one color in [LineChartBarData.mainColors], it returns that color in a darker mode,
 /// otherwise it returns the color along the gradient colors based on the [xPercentage] in a darker mode.
-Color _defaultGetDotStrokeColor(
-    FlSpot spot, double xPercentage, LineChartBarData bar) {
+Color _defaultGetDotStrokeColor(FlSpot spot, double xPercentage, LineChartBarData bar) {
   Color color;
   if (bar.gradient != null && bar.gradient is LinearGradient) {
     color = lerpGradient(
@@ -769,12 +735,9 @@ Color _defaultGetDotStrokeColor(
 /// [LineChartBarData] is the chart's bar.
 /// [int] is the index position of the spot.
 /// It should return a [FlDotPainter] that needs to be used for drawing target.
-typedef GetDotPainterCallback = FlDotPainter Function(
-    FlSpot, double, LineChartBarData, int);
+typedef GetDotPainterCallback = FlDotPainter Function(FlSpot, double, LineChartBarData, int);
 
-FlDotPainter _defaultGetDotPainter(
-    FlSpot spot, double xPercentage, LineChartBarData bar, int index,
-    {double? size}) {
+FlDotPainter _defaultGetDotPainter(FlSpot spot, double xPercentage, LineChartBarData bar, int index, {double? size}) {
   return FlDotCirclePainter(
     radius: size,
     color: _defaultGetDotColor(spot, xPercentage, bar),
@@ -1035,140 +998,6 @@ bool showAllDots(FlSpot spot, LineChartBarData barData) {
   return true;
 }
 
-/// Holds data for drawing extra horizontal lines.
-///
-/// [LineChart] draws some [HorizontalLine] (set by [LineChartData.extraLinesData]),
-/// in below or above of everything, it draws from left to right side of the chart.
-class HorizontalLine extends FlLine with EquatableMixin {
-  /// Draws from left to right of the chart using the [y] value.
-  final double y;
-
-  /// Use it for any kind of image, to draw it in left side of the chart.
-  Image? image;
-
-  /// Use it for vector images, to draw it in left side of the chart.
-  SizedPicture? sizedPicture;
-
-  /// Draws a text label over the line.
-  final HorizontalLineLabel label;
-
-  /// [LineChart] draws horizontal lines from left to right side of the chart
-  /// in the provided [y] value, and color it using [color].
-  /// You can define the thickness using [strokeWidth]
-  ///
-  /// It draws a [label] over it.
-  ///
-  /// You can have a dashed line by filling [dashArray] with dash size and space respectively.
-  ///
-  /// It draws an image in left side of the chart, use [sizedPicture] for vectors,
-  /// or [image] for any kind of image.
-  HorizontalLine({
-    required this.y,
-    HorizontalLineLabel? label,
-    Color? color,
-    double? strokeWidth,
-    List<int>? dashArray,
-    this.image,
-    this.sizedPicture,
-  })  : label = label ?? HorizontalLineLabel(),
-        super(
-            color: color ?? Colors.black,
-            strokeWidth: strokeWidth ?? 2,
-            dashArray: dashArray);
-
-  /// Lerps a [HorizontalLine] based on [t] value, check [Tween.lerp].
-  static HorizontalLine lerp(HorizontalLine a, HorizontalLine b, double t) {
-    return HorizontalLine(
-      y: lerpDouble(a.y, b.y, t)!,
-      label: HorizontalLineLabel.lerp(a.label, b.label, t),
-      color: Color.lerp(a.color, b.color, t),
-      strokeWidth: lerpDouble(a.strokeWidth, b.strokeWidth, t),
-      dashArray: lerpIntList(a.dashArray, b.dashArray, t),
-      image: b.image,
-      sizedPicture: b.sizedPicture,
-    );
-  }
-
-  /// Used for equality check, see [EquatableMixin].
-  @override
-  List<Object?> get props => [
-        y,
-        label,
-        color,
-        strokeWidth,
-        dashArray,
-        image,
-        sizedPicture,
-      ];
-}
-
-/// Holds data for drawing extra vertical lines.
-///
-/// [LineChart] draws some [VerticalLine] (set by [LineChartData.extraLinesData]),
-/// in below or above of everything, it draws from bottom to top side of the chart.
-class VerticalLine extends FlLine with EquatableMixin {
-  /// Draws from bottom to top of the chart using the [x] value.
-  final double x;
-
-  /// Use it for any kind of image, to draw it in bottom side of the chart.
-  Image? image;
-
-  /// Use it for vector images, to draw it in bottom side of the chart.
-  SizedPicture? sizedPicture;
-
-  /// Draws a text label over the line.
-  final VerticalLineLabel label;
-
-  /// [LineChart] draws vertical lines from bottom to top side of the chart
-  /// in the provided [x] value, and color it using [color].
-  /// You can define the thickness using [strokeWidth]
-  ///
-  /// It draws a [label] over it.
-  ///
-  /// You can have a dashed line by filling [dashArray] with dash size and space respectively.
-  ///
-  /// It draws an image in bottom side of the chart, use [sizedPicture] for vectors,
-  /// or [image] for any kind of image.
-  VerticalLine({
-    required this.x,
-    VerticalLineLabel? label,
-    Color? color,
-    double? strokeWidth,
-    List<int>? dashArray,
-    this.image,
-    this.sizedPicture,
-  })  : label = label ?? VerticalLineLabel(),
-        super(
-            color: color ?? Colors.black,
-            strokeWidth: strokeWidth ?? 2,
-            dashArray: dashArray);
-
-  /// Lerps a [VerticalLine] based on [t] value, check [Tween.lerp].
-  static VerticalLine lerp(VerticalLine a, VerticalLine b, double t) {
-    return VerticalLine(
-      x: lerpDouble(a.x, b.x, t)!,
-      label: VerticalLineLabel.lerp(a.label, b.label, t),
-      color: Color.lerp(a.color, b.color, t),
-      strokeWidth: lerpDouble(a.strokeWidth, b.strokeWidth, t),
-      dashArray: lerpIntList(a.dashArray, b.dashArray, t),
-      image: b.image,
-      sizedPicture: b.sizedPicture,
-    );
-  }
-
-  /// Used for equality check, see [EquatableMixin].
-  @override
-  List<Object?> get props => [
-        x,
-        label,
-        color,
-        strokeWidth,
-        dashArray,
-        image,
-        sizedPicture,
-      ];
-}
-
 /// Shows a text label
 abstract class FlLineLabel with EquatableMixin {
   /// Determines showing label or not.
@@ -1187,11 +1016,7 @@ abstract class FlLineLabel with EquatableMixin {
   /// applies [padding] for spaces, and applies [style] for changing color,
   /// size, ... of the text.
   /// [show] determines showing label or not.
-  FlLineLabel(
-      {required this.show,
-      required this.padding,
-      required this.style,
-      required this.alignment});
+  FlLineLabel({required this.show, required this.padding, required this.style, required this.alignment});
 
   /// Used for equality check, see [EquatableMixin].
   @override
@@ -1209,8 +1034,7 @@ class HorizontalLineLabel extends FlLineLabel with EquatableMixin {
   final String Function(HorizontalLine) labelResolver;
 
   /// Returns the [HorizontalLine.y] as the drawing label.
-  static String defaultLineLabelResolver(HorizontalLine line) =>
-      line.y.toStringAsFixed(1);
+  static String defaultLineLabelResolver(HorizontalLine line) => line.y.toStringAsFixed(1);
 
   /// Draws a title on the [HorizontalLine], align it with [alignment] over the line,
   /// applies [padding] for spaces, and applies [style for changing color,
@@ -1224,8 +1048,7 @@ class HorizontalLineLabel extends FlLineLabel with EquatableMixin {
     Alignment? alignment,
     bool show = false,
     String Function(HorizontalLine)? labelResolver,
-  })  : labelResolver =
-            labelResolver ?? HorizontalLineLabel.defaultLineLabelResolver,
+  })  : labelResolver = labelResolver ?? HorizontalLineLabel.defaultLineLabelResolver,
         super(
           show: show,
           padding: padding ?? const EdgeInsets.all(6),
@@ -1234,11 +1057,9 @@ class HorizontalLineLabel extends FlLineLabel with EquatableMixin {
         );
 
   /// Lerps a [HorizontalLineLabel] based on [t] value, check [Tween.lerp].
-  static HorizontalLineLabel lerp(
-      HorizontalLineLabel a, HorizontalLineLabel b, double t) {
+  static HorizontalLineLabel lerp(HorizontalLineLabel a, HorizontalLineLabel b, double t) {
     return HorizontalLineLabel(
-      padding:
-          EdgeInsets.lerp(a.padding as EdgeInsets, b.padding as EdgeInsets, t),
+      padding: EdgeInsets.lerp(a.padding as EdgeInsets, b.padding as EdgeInsets, t),
       style: TextStyle.lerp(a.style, b.style, t),
       alignment: Alignment.lerp(a.alignment, b.alignment, t),
       labelResolver: b.labelResolver,
@@ -1263,8 +1084,7 @@ class VerticalLineLabel extends FlLineLabel with EquatableMixin {
   final String Function(VerticalLine) labelResolver;
 
   /// Returns the [VerticalLine.x] as the drawing label.
-  static String defaultLineLabelResolver(VerticalLine line) =>
-      line.x.toStringAsFixed(1);
+  static String defaultLineLabelResolver(VerticalLine line) => line.x.toStringAsFixed(1);
 
   /// Draws a title on the [VerticalLine], align it with [alignment] over the line,
   /// applies [padding] for spaces, and applies [style for changing color,
@@ -1278,8 +1098,7 @@ class VerticalLineLabel extends FlLineLabel with EquatableMixin {
     Alignment? alignment,
     bool? show,
     String Function(VerticalLine)? labelResolver,
-  })  : labelResolver =
-            labelResolver ?? VerticalLineLabel.defaultLineLabelResolver,
+  })  : labelResolver = labelResolver ?? VerticalLineLabel.defaultLineLabelResolver,
         super(
           show: show ?? false,
           padding: padding ?? const EdgeInsets.all(6),
@@ -1293,11 +1112,9 @@ class VerticalLineLabel extends FlLineLabel with EquatableMixin {
         );
 
   /// Lerps a [VerticalLineLabel] based on [t] value, check [Tween.lerp].
-  static VerticalLineLabel lerp(
-      VerticalLineLabel a, VerticalLineLabel b, double t) {
+  static VerticalLineLabel lerp(VerticalLineLabel a, VerticalLineLabel b, double t) {
     return VerticalLineLabel(
-      padding:
-          EdgeInsets.lerp(a.padding as EdgeInsets, b.padding as EdgeInsets, t),
+      padding: EdgeInsets.lerp(a.padding as EdgeInsets, b.padding as EdgeInsets, t),
       style: TextStyle.lerp(a.style, b.style, t),
       alignment: Alignment.lerp(a.alignment, b.alignment, t),
       labelResolver: b.labelResolver,
@@ -1354,47 +1171,6 @@ class SizedPicture with EquatableMixin {
         picture,
         width,
         height,
-      ];
-}
-
-/// Draws some straight horizontal or vertical lines in the [LineChart]
-class ExtraLinesData with EquatableMixin {
-  final List<HorizontalLine> horizontalLines;
-  final List<VerticalLine> verticalLines;
-
-  final bool extraLinesOnTop;
-
-  /// [LineChart] draws some straight horizontal or vertical lines,
-  /// you should set [LineChartData.extraLinesData].
-  /// Draws horizontal lines using [horizontalLines],
-  /// and vertical lines using [verticalLines].
-  ///
-  /// If [extraLinesOnTop] sets true, it draws the line above the main bar lines, otherwise
-  /// it draws them below the main bar lines.
-  ExtraLinesData({
-    List<HorizontalLine>? horizontalLines,
-    List<VerticalLine>? verticalLines,
-    bool? extraLinesOnTop,
-  })  : horizontalLines = horizontalLines ?? const [],
-        verticalLines = verticalLines ?? const [],
-        extraLinesOnTop = extraLinesOnTop ?? true;
-
-  /// Lerps a [ExtraLinesData] based on [t] value, check [Tween.lerp].
-  static ExtraLinesData lerp(ExtraLinesData a, ExtraLinesData b, double t) {
-    return ExtraLinesData(
-      extraLinesOnTop: b.extraLinesOnTop,
-      horizontalLines:
-          lerpHorizontalLineList(a.horizontalLines, b.horizontalLines, t),
-      verticalLines: lerpVerticalLineList(a.verticalLines, b.verticalLines, t),
-    );
-  }
-
-  /// Used for equality check, see [EquatableMixin].
-  @override
-  List<Object?> get props => [
-        horizontalLines,
-        verticalLines,
-        extraLinesOnTop,
       ];
 }
 
@@ -1457,8 +1233,7 @@ class LineTouchData extends FlTouchData<LineTouchResponse> with EquatableMixin {
     GetTouchLineY? getTouchLineStart,
     GetTouchLineY? getTouchLineEnd,
   })  : touchTooltipData = touchTooltipData ?? LineTouchTooltipData(),
-        getTouchedSpotIndicator =
-            getTouchedSpotIndicator ?? defaultTouchedIndicators,
+        getTouchedSpotIndicator = getTouchedSpotIndicator ?? defaultTouchedIndicators,
         touchSpotThreshold = touchSpotThreshold ?? 10,
         distanceCalculator = distanceCalculator ?? _xDistance,
         handleBuiltInTouches = handleBuiltInTouches ?? true,
@@ -1485,8 +1260,7 @@ class LineTouchData extends FlTouchData<LineTouchResponse> with EquatableMixin {
       touchCallback: touchCallback ?? this.touchCallback,
       mouseCursorResolver: mouseCursorResolver ?? this.mouseCursorResolver,
       touchTooltipData: touchTooltipData ?? this.touchTooltipData,
-      getTouchedSpotIndicator:
-          getTouchedSpotIndicator ?? this.getTouchedSpotIndicator,
+      getTouchedSpotIndicator: getTouchedSpotIndicator ?? this.getTouchedSpotIndicator,
       touchSpotThreshold: touchSpotThreshold ?? this.touchSpotThreshold,
       distanceCalculator: distanceCalculator ?? this.distanceCalculator,
       getTouchLineStart: getTouchLineStart ?? this.getTouchLineStart,
@@ -1521,12 +1295,10 @@ typedef GetTouchedSpotIndicator = List<TouchedSpotIndicatorData?> Function(
     LineChartBarData barData, List<int> spotIndexes);
 
 /// Used for determine the touch indicator line's starting/end point.
-typedef GetTouchLineY = double Function(
-    LineChartBarData barData, int spotIndex);
+typedef GetTouchLineY = double Function(LineChartBarData barData, int spotIndex);
 
 /// Used to calculate the distance between coordinates of a touch event and a spot
-typedef CalculateTouchDistance = double Function(
-    Offset touchPoint, Offset spotPixelCoordinates);
+typedef CalculateTouchDistance = double Function(Offset touchPoint, Offset spotPixelCoordinates);
 
 /// Default distanceCalculator only considers distance on x axis
 double _xDistance(Offset touchPoint, Offset spotPixelCoordinates) {
@@ -1534,8 +1306,7 @@ double _xDistance(Offset touchPoint, Offset spotPixelCoordinates) {
 }
 
 /// Default presentation of touched indicators.
-List<TouchedSpotIndicatorData> defaultTouchedIndicators(
-    LineChartBarData barData, List<int> indicators) {
+List<TouchedSpotIndicatorData> defaultTouchedIndicators(LineChartBarData barData, List<int> indicators) {
   return indicators.map((int index) {
     /// Indicator Line
     var lineColor = barData.gradient?.colors.first ?? barData.color;
@@ -1551,8 +1322,7 @@ List<TouchedSpotIndicatorData> defaultTouchedIndicators(
     }
 
     final dotData = FlDotData(
-        getDotPainter: (spot, percent, bar, index) =>
-            _defaultGetDotPainter(spot, percent, bar, index, size: dotSize));
+        getDotPainter: (spot, percent, bar, index) => _defaultGetDotPainter(spot, percent, bar, index, size: dotSize));
 
     return TouchedSpotIndicatorData(flLine, dotData);
   }).toList();
@@ -1630,8 +1400,7 @@ class LineTouchTooltipData with EquatableMixin {
     BorderSide? tooltipBorder,
   })  : tooltipBgColor = tooltipBgColor ?? Colors.blueGrey.darken(15),
         tooltipRoundedRadius = tooltipRoundedRadius ?? 4,
-        tooltipPadding = tooltipPadding ??
-            const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+        tooltipPadding = tooltipPadding ?? const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
         tooltipMargin = tooltipMargin ?? 16,
         maxContentWidth = maxContentWidth ?? 120,
         getTooltipItems = getTooltipItems ?? defaultLineTooltipItem,
@@ -1666,16 +1435,13 @@ class LineTouchTooltipData with EquatableMixin {
 /// then you should and pass your custom [LineTooltipItem] list
 /// (length should be equal to the [touchedSpots.length]),
 /// to show inside the tooltip popup.
-typedef GetLineTooltipItems = List<LineTooltipItem?> Function(
-    List<LineBarSpot> touchedSpots);
+typedef GetLineTooltipItems = List<LineTooltipItem?> Function(List<LineBarSpot> touchedSpots);
 
 /// Default implementation for [LineTouchTooltipData.getTooltipItems].
 List<LineTooltipItem> defaultLineTooltipItem(List<LineBarSpot> touchedSpots) {
   return touchedSpots.map((LineBarSpot touchedSpot) {
     final textStyle = TextStyle(
-      color: touchedSpot.bar.gradient?.colors.first ??
-          touchedSpot.bar.color ??
-          Colors.blueGrey,
+      color: touchedSpot.bar.gradient?.colors.first ?? touchedSpot.bar.color ?? Colors.blueGrey,
       fontWeight: FontWeight.bold,
       fontSize: 14,
     );
@@ -1799,8 +1565,7 @@ class ShowingTooltipIndicators with EquatableMixin {
 
   /// [LineChart] shows some tooltips over each [LineChartBarData],
   /// and [showingSpots] determines in which spots this tooltip should be shown.
-  ShowingTooltipIndicators(List<LineBarSpot> showingSpots)
-      : showingSpots = showingSpots;
+  ShowingTooltipIndicators(List<LineBarSpot> showingSpots) : showingSpots = showingSpots;
 
   /// Used for equality check, see [EquatableMixin].
   @override
@@ -1834,8 +1599,7 @@ class LineTouchResponse extends BaseTouchResponse {
 
 /// It lerps a [LineChartData] to another [LineChartData] (handles animation for updating values)
 class LineChartDataTween extends Tween<LineChartData> {
-  LineChartDataTween({required LineChartData begin, required LineChartData end})
-      : super(begin: begin, end: end);
+  LineChartDataTween({required LineChartData begin, required LineChartData end}) : super(begin: begin, end: end);
 
   /// Lerps a [LineChartData] based on [t] value, check [Tween.lerp].
   @override