+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * **BUGFIX** (by @ateich): Fix infinite loop in RadarChart when all values in RadarDataSet are equal, #882.
 * **BUGFIX** (by @ateich): Fix uneven titles in RadarChart when using titlePositionPercentageOffset, #1074.
 * **BUGFIX** (by @imaNNeoFighT): Fix PieChart single section stroke issue, #1089
+* **FEATURE** Now we can use ExtraLinesData for LineChart and BarChart, #476
 
 ## 0.55.0
 * **FEATURE** (by @emelinepal): Add `tooltipBorder` property in [LineTouchTooltipData], [BarTouchTooltipData], [ScatterTouchTooltipData], #692.